Ensures that spree db:remigrate task resets cached db schema info
This should be reported as a bug to the Spree core team. The migration
that creates the spree_tokenized_permissions table first asks the
table's ActiveRecord model if the table exists. However, during
the remigrate task, the tables are dropped right before migrations
are run. This prevents the spree_tokenized_permissions table from
getting created. That will then prevent our migration to add a column
to that table to fail.

The work around is to ensure that the cached schema data gets reset
before the db:migrate task is run.